requirements.txt
streamlit
pillow
numpy
pygame
